+++ src/core/io_wrapper.h
